🐛 (engine) Load factions only if faction selected
The faction variant is stored and loaded from the json. This will cause changes in the balanced set to be pinned for the whole game. However, when bidding, the faction can change owner. The previous player's faction is unset. By moving the load of the faction variant into the if-statement, the cached variant is discarded after switching owners.