🏗️ (all) Change the way faction versions are stored
Now the "version" field is required for variants. When the engine is created, the factionVersion field is automatically set to the latest version for the variant.

When a player choose their faction, the faction customization is stored, with the version (not used, just metadata) and now even handlers from the variant are saved (but they can't use local scope for now, they need to be pure functions).

In the engine, apart from the faction variant name & version number, nothing is stored.

During a replay, the player will load its own faction first, if not possible then the matching faction corresponding to engine's faction variant / version.